fix(random): PCG64 Generator parity audit vs NumPy 2.4.2 — 6 byte-exact fixes
Full differential audit of the modern PCG64 Generator subsystem against NumPy
2.4.2 (~1,100 differential cases across seeds/dtypes/sizes/layouts/params, plus
error-message and out= contracts). Found and fixed 6 divergences, each verified
byte-exact with zero regressions (482 RandomSampling unit + 107 oracle cases green):

1. float32 SCALAR returned a float32 scalar; NumPy returns float64. For size=None,
   _common.pyx float_fill does 'random_func(&out_val); return out_val', so the C
   float auto-converts to a Python float (float64) — only sized/out= stay float32.
   Fixed the size=None branch of random / standard_normal / standard_exponential /
   standard_gamma to widen the float32 draw to a float64 scalar.

2. ValidateOut used the wrong message and check order. Now mirrors NumPy's
   check_output (require_c_array=False): contiguity -> dtype -> shape, the verbatim
   'Supplied output array must be contiguous, writable, aligned, and in machine
   byte-order.' message, and accepts a C- OR F-contiguous writable out.

3. choice(replace=False, p=weights) threw NotSupportedException — NumPy supports it.
   Ported the weighted-without-replacement algorithm (ChoiceNoReplaceWeighted): per
   round draw random((size-n_uniq,)), zero found probabilities, cumsum -> /cdf[-1] ->
   bisect_right(side='right'), first-occurrence dedup; byte-exact incl. the
   'Fewer non-zero entries in p than size' guard.

4. shuffle on a 0-d array raised ArgumentException; NumPy evaluates n=len(x) first, so
   it raises TypeError('len() of unsized object'). Added the ndim==0 guard up front.

5. The uint64 range above long.MaxValue was unreachable (integers' high is 'long').
   Added a ulong overload integers(ulong, ulong?, ...) that forwards in-range values
   to the signed path and handles the genuine uint64-large case; the full [0, 2**64)
   range is drawn as integers(0UL, ulong.MaxValue, dtype: np.uint64, endpoint: true).

6. permuted was entirely missing. Implemented it (axis=None shuffles the C-order flat
   copy; an explicit axis runs an independent random_interval Fisher-Yates over each
   1-D slice, iterating the remaining axes in C-order per NumPy's IterAllButAxis; out=
   is in-place and returned). Byte-exact incl. 3-D and negative-axis cases.

Pins: test/NumSharp.Tests/RandomSampling/Generator.ParityFixes.Test.cs (16 tests).

Remaining gaps (feature-scale additions, not defects — a C# caller cannot express
them today, so there is no wrong output): array-valued distribution parameters
(loc/scale/low/high/shape broadcasting) for uniform/normal/exponential/gamma/
standard_gamma and integers array low/high; and the parallel-stream infra
PCG64.advance/jumped and SeedSequence.spawn/spawn_key/pool_size.

+++ src/NumSharp.Core/RandomSampling/Generator.Exponential.cs
@@ -32,7 +32,9 @@ namespace NumSharp
             }
 
             if (IsNoSize(size))
-                return tc == NPTypeCode.Single ? NDArray.Scalar(f()) : NDArray.Scalar(d());
+                // size=None returns a float64 scalar even for dtype=float32 (NumPy float_fill widens
+                // the float32 draw to a Python float); only sized/out= stay float32.
+                return tc == NPTypeCode.Single ? NDArray.Scalar((double)f()) : NDArray.Scalar(d());
 
             return tc == NPTypeCode.Single ? FillFloatDist(size, f) : FillDoubleDist(size, d);
         }
